# improve gcd performance